ENHANCE(): Arithmetic logic

+++ app/src/main/java/com/dhandev/rekapin/data/model/ProfileModel.kt
@@ -2,8 +2,8 @@ package com.dhandev.rekapin.data.model
 
 data class ProfileModel(
     var userName: String,
-    var balance: Long,
-    val budget: Long,
-    val target: Long,
+    var balance: Double,
+    val budget: Double,
+    val target: Double,
     val reportPeriod: Int
 )
